test(alert): move htmlAttributes id test to alert-id.spec.tsx
